s2

+++ Service2/Controllers/MoneyController.cs
@@ -1,7 +1,11 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Caching.Distributed;
 using MongoDB.Driver;
-using Service.Models;
 using Prometheus;
+using Service.Models;
+using StackExchange.Redis;
+using System.Text.Json;
+
 
 namespace WebAPIApp.Controllers
 {
@@ -10,6 +14,8 @@ namespace WebAPIApp.Controllers
     public class MoneyController : ControllerBase
     {
         private readonly IMongoCollection<Money> _moneyCollection;
+        private readonly IDistributedCache _cache;
+       
         private static readonly Counter RequestsTotal =
     Metrics.CreateCounter("api_requests_total2", "");
         private static readonly Gauge MoneyCount =
@@ -31,9 +37,9 @@ namespace WebAPIApp.Controllers
                 });
 
 
-        public MoneyController()
+        public MoneyController(IDistributedCache cache)
         {
-
+            _cache = cache;
             var client = new MongoClient("mongodb://localhost:27017");
             var database = client.GetDatabase("Numismatic_Club2");
             _moneyCollection = database.GetCollection<Money>("Money");
@@ -47,10 +53,17 @@ namespace WebAPIApp.Controllers
 
         // GET api/money
         [HttpGet]
-
         public async Task<ActionResult<Money>> Get()
         {
             RequestsTotal.Inc();
+            string cacheKey = "money_all";
+
+            var cached = await _cache.GetStringAsync(cacheKey);
+            if (cached != null)
+            {
+                var data = JsonSerializer.Deserialize<object>(cached);
+                return Ok(data);
+            }
             var totalCount = await _moneyCollection.CountDocumentsAsync(_ => true);
 
             List<Money> entities;
@@ -75,21 +88,38 @@ namespace WebAPIApp.Controllers
                 m.Year_of_creation,
                 m.Country
             });
+            var options = new DistributedCacheEntryOptions
+            {
+                AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+            };
+
+
+            await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
 
             return Ok(result);
         }
 
         // GET api/money/{id}
         [HttpGet("{id}")]
-
-
-
         public async Task<ActionResult<Money_>> Get(string id)
         {
             RequestsTotal.Inc();
 
-            using (RequestLatency.NewTimer())
+
             {
+                if (id.Length != 24)
+                {
+                    ErrorsTotal2.Inc();
+                    return NotFound("id состоит из 24 символов");
+                }
+                string cacheKey = $"money_{id}";
+
+                var cached = await _cache.GetStringAsync(cacheKey);
+                if (cached != null)
+                {
+                    var data = JsonSerializer.Deserialize<Money_>(cached);
+                    return Ok(data);
+                }
                 var money = await _moneyCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
 
                 if (money == null)
@@ -97,15 +127,17 @@ namespace WebAPIApp.Controllers
                     ErrorsTotal2.Inc();
                     return NotFound();
                 }
-
-
                 var result = new Money_
                 {
                     Name = money.Name,
                     Year_of_creation = money.Year_of_creation,
                     Country = money.Country
                 };
-
+                var options = new DistributedCacheEntryOptions
+                {
+                    AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+                };
+                await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
                 return Ok(result);
             }
         }
@@ -117,10 +149,11 @@ namespace WebAPIApp.Controllers
         {
             RequestsTotal.Inc();
 
-            using (RequestLatency.NewTimer())
+
             {
                 try
                 {
+
                     if (moneyWithoutId == null)
                     {
                         ErrorsTotal2.Inc();
@@ -162,12 +195,13 @@ namespace WebAPIApp.Controllers
                         Year_of_creation = money.Year_of_creation,
                         Country = money.Country
                     };
+                    await _cache.RemoveAsync("money_all");
                     MoneyCount.Inc();
                     return CreatedAtAction(nameof(Get), new { id = money.Id }, result);
                 }
                 catch
                 {
-                    ErrorsTotal2.Inc();
+
                     throw;
                 }
             }
@@ -178,20 +212,27 @@ namespace WebAPIApp.Controllers
         public async Task<ActionResult> Put(string id, [FromBody] Money_ moneyWithoutId)
 
         {
-            RequestsTotal.Inc();
 
-            using (RequestLatency.NewTimer())
+
+            RequestsTotal.Inc();
+            if (id.Length != 24)
+            {
+                ErrorsTotal2.Inc();
+                return NotFound("id состоит из 24 символов");
+            }
+
             {
                 if (moneyWithoutId == null)
                 {
-                    ErrorsTotal2.Inc();
+
                     return BadRequest();
                 }
+
 
                 var existingMoney = await _moneyCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
                 if (existingMoney == null)
                 {
-                    ErrorsTotal2.Inc();
+
                     return NotFound();
                 }
 
@@ -208,10 +249,12 @@ namespace WebAPIApp.Controllers
                     x => x.Id == id,
                     updatedMoney
                 );
+                await _cache.RemoveAsync("money_all");
+                await _cache.RemoveAsync($"money_{id}");
 
                 if (result.MatchedCount == 0)
                 {
-                    ErrorsTotal2.Inc();
+
                     return NotFound();
                 }
 
@@ -233,6 +276,12 @@ namespace WebAPIApp.Controllers
         {
             RequestsTotal.Inc();
 
+            if (id.Length != 24)
+            {
+                ErrorsTotal2.Inc();
+                return NotFound("id состоит из 24 символов");
+            }
+
             var result = await _moneyCollection.DeleteOneAsync(x => x.Id == id);
 
             if (result.DeletedCount == 0)
@@ -240,6 +289,11 @@ namespace WebAPIApp.Controllers
                 ErrorsTotal2.Inc();
                 return NotFound();
             }
+
+
+
+            await _cache.RemoveAsync($"money_{id}");
+            await _cache.RemoveAsync("money_all");
             MoneyCount.Dec();
             return Ok();
         }
@@ -248,10 +302,16 @@ namespace WebAPIApp.Controllers
         [HttpGet("filter")]
         public async Task<ActionResult<Money_>> Filter([FromQuery] int? year, [FromQuery] string comparison)
         {
-            RequestsTotal.Inc();
             var query = _moneyCollection.AsQueryable();
+            RequestsTotal.Inc();
+            string cacheKey = $"money_filter_{comparison}_{year}";
 
-
+            var cached = await _cache.GetStringAsync(cacheKey);
+            if (cached != null)
+            {
+                var data = JsonSerializer.Deserialize<IEnumerable<Money_>>(cached);
+                return Ok(data);
+            }
             switch (comparison.ToLower())
             {
                 case "больше":
@@ -275,7 +335,11 @@ namespace WebAPIApp.Controllers
                 Year_of_creation = m.Year_of_creation,
                 Country = m.Country
             });
-
+            var options = new DistributedCacheEntryOptions
+            {
+                AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(5)
+            };
+            await _cache.SetStringAsync(cacheKey, JsonSerializer.Serialize(result), options);
             return Ok(result);
         }
     }
